Create PhoneKeypadCombinations_in_CPP
Given a sequence of digits (0-9) representing a phone keypad, generate and print all possible combinations of characters that can be formed by typing the given sequence on the keypad. Each digit is associated with a set of characters, and the program should explore all combinations of those characters for the given input sequence.